Fixed the league deserialisation issue and made the normal Elo thing faster, should probably revert it, though

+++ LibOfLegends/com/riotgames/leagues/pojo/MiniSeriesDTO.cs
@@ -6,7 +6,7 @@ namespace com.riotgames.leagues.pojo
 	{
 		public int target;
 		public int losses;
-		public int timeLeftToPlayMillis;
+		public long timeLeftToPlayMillis;
 		public int wins;
 	}
 }
